Add command summary to README and fix incorrect `ekg-notes-tag-any'
We previously documented that `ekg-notes-tag-any' will expand to all the tags
in the buffer.  However, that's not what it did, it only expanded to tht tags in
the current note.

Because of that, we renamed `ekg-notes-any-tags' to `ekg-notes-any-note-tags' to
encompass the previous behavior, and created a new function `ekg-notes-any-tags'
to execute the behavior as documented. Bindings are updated so that "A" is the
previously documented behavior.