Remove separate packaging for ekg-org-roam
Now that ekg-org-roam has been added to the ekg package on MELPA, there is no
need for separate package stanzas in ekg-org-roam.el.

Change the README to adapt to the new packaging setup.

In ekg-org-roam.el, use `split-string' consistently, instead of the newer
`string-split', for compatibility with older emacs versions.  The two functions
are identical.