Improve trash viewing
Fix the trash note buffer, giving it a reasonable name and calculating it
in a more straighforward manner.  Remove the notion of seeing just anything with
a trashed tag if not everything is trashed.  This may be useful, but I want to
rethink if trashing individual tags and not everything is a concept we want to keep.