Don't save text propertized
This is an important bug fix, incrementing the version to 0.1.1.  I noticed that
sometimes propertized strings, for a reason I haven't figured out yet, cannot be
read back out of the database.  It's important to avoid anything that might
corrupt the database.

Also, the properties really blow up the size of the database, for no good
reason, since we always display the font-locked version of the text in the
proper mode.

The only use for propertized strings might be in text-mode, but unless someone
wants it specifically, let's just stop doing for now.