Change metadata section to have after-string, resist deletion
The metadata section now removes the string that read "Note Properties", and now
has "--text follows this line--", which is exactly way message-mode presents the
distinction between their equivalent of metadata properties and the text.

Also, add new tests to make sure that deleting characters at the end of the
overlay does not harm the overlay.