Change so that the line after the metadata cannot be modified
It used to be that deleting from this line would delete the metadata, or adding
to it would add things to the metadata, and both were a bit weird.

Unfortunately, this breaks (or perhaps it was broken before) the ability to use
`org-insert-heading', which doesn't work well with our metadata. This seems
unfortunate and probably something needing further fixes.

This fixes https://github.com/ahyatt/ekg/issues/43.