Make sure resource removal doesn't result in empty ids
We now make sure to always have an ID, and if it doesn't exist, generate one.
This has to be done before renaming the resource, but also in
`ekg--normalize-note', for any other client who is saving a note, but not via
editing a note.

Also change the renaming to right before saving the note, and include it in the
same transaction.

This fixes https://github.com/ahyatt/ekg/issues/46.