Remove threading from ekg-embedding-generate-all, fix count
Threading never really works the way I'd like it to, perhaps I'm not using it
well.  The idea is that by yielding the thread, everything is responsive for the
user, but I'd expect the thread to be yielded back all the time during any
pauses.  That didn't seem to happen.

The count was previously counting all embeddings that were checked, not all
embeddings generated. Now we just inform the user of the number of generated
embeddings.