Store ekg version, use to perform upgrade, even when ekg is connected
Change `ekg-connect' so that the check to see if it upgrades is independent of
the connection state.  That way, if the user has a connection open, but upgrades
ekg, it will do the right thing and upgrade the schema and any other upgrades.

This also fixes test that previously used the obsolete `ekg--connect', and also
needed to set the db version so it wouldn't be upgraded (which may not hurt but
it didn't make sense for tests).

Also, let's finally remove a few upgrades that probably don't need to be run
anymore, because anyone who was running on these older versions has already
upgraded.  I think.