Switch to a different method of handling module schema
Previously, we relied on the ekg-add-schema-hook, but with recent changes, that
will no longer be run on every connection.  It also wasn't reliably run anyway,
in cases where the connection was open but the user was using the module (such
as ekg-logseq), for the first time.

Now, each module has their own connection method, which they use at appropriate
places to make sure the connection is open and their particular schema is in
place.

This fixes the underlying issue behind https://github.com/ahyatt/ekg/issues/62.