Add magic tags, which run elisp from a co-tagged note
The default tag, "tag-defun", will combine with another tag in exactly the same
way templates work, but instead of working only when added, it will be executed
when the note is loaded or the tag is added. Note that the tag currently has to
be added with completion to take effect.