Add the ability to saved "drafts"
Notes can be saved with tag "drafts", and there's a new function,
ekg-show-notes-in-drafts that can show them.  The ordinary buffer save command
is remapped to save drafts.

This basic approach was proposed by http://github.com/qingshuizheng in
https://github.com/ahyatt/ekg/discussions/80, and a separate solution, without
the tagging approach was given by https://github.com/PuercoPop in
https://github.com/ahyatt/ekg/pull/19.  Thank you to both for your help on
getting this functionality out.