Fix and make embedding tag updating more robust
There were a few things wrong, first of all, the tag updates that were done
before saving a note did nothing, they need to be done after the saving of the
note.

Also, fix ekg-embedding-valid-p, which was not working properly before.

When updating tags, fix any invalid embeddings found along the way.

Ignore any errors when updatings tags; these are not so important that they need
to be immediately surfaced to the user as an error.