Fix issue with tags being duplicated after a global rename
The triple store doesn't have a way to uniquify lists at the moment, and even if
it did, we don't use it directly to rename tags. That might be useful to add as
a feature on the triple store, but for now, let's just fix this directly.

Also adds cleanup for this in ekg-clean-db.

This fixes https://github.com/ahyatt/ekg/discussions/111.