Fix issue where titles were not getting removed for deleted notes
In general, this is a hard problem to solve - this current solution is
temporary. What really needs to happen (I think) is that the triples package
needs to encode type dependency information, which it then can use automatically
to make sure various things are deleted when necessary.

This fixes https://github.com/ahyatt/ekg/issues/112.