Refine fix for auto-narrowing to just org-insert commands
Because so many org commands are rebindings of normal movement commands, this is
necessary to keep basic movement working.  It's likely we'll have to expand this
to more substrings of org commands, but for now org-insert is a likely culprit.

This fixes issues with the fix in https://github.com/ahyatt/ekg/issues/121.