Simplify commands and actions on buffer killing
Remove ekg-save-action-on-buffer-kill (never in a release yet), and
ekg-confirm-on-abort.  Add ekg-confirm-on-buffer-kill.  The thinking is that the
user can just choose not to kill the buffer if they choose, and then they can
save, or abort, or do whatever they want after that.

This also removes the possibility of aborting on editing, which is confusing and
I'm still not sure I understood what it did.